requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
# Core packages only - minimize dependencies
openai==1.12.0
anthropic==0.8.0
# Simple JSON-based operations only, no data science
# langchain-core and langgraph are too heavy
# pandas and numpy are too heavy
